Moved some fixtures out of test_easy_install.
