+++ setuptools/tests/test_dist.py
@@ -8,7 +8,7 @@ import pytest
 from setuptools import Distribution
 from setuptools.dist import check_package_data, check_specifier
 
-from .test_easy_install import make_trivial_sdist
+from .fixtures import make_trivial_sdist
 from .test_find_packages import ensure_files
 from .textwrap import DALS
