+++ setuptools/tests/test_easy_install.py
@@ -1,15 +1,12 @@
 """Easy install Tests"""
 
-import io
 import itertools
 import logging
 import os
 import re
 import site
 import sys
-import tarfile
 import tempfile
-import time
 import warnings
 import zipfile
 from typing import NamedTuple
@@ -20,10 +17,10 @@ import pytest
 import pkg_resources
 import setuptools.command.easy_install as ei
 from pkg_resources import Distribution as PRDistribution, normalize_path
-from setuptools._normalization import safer_name
 from setuptools.command.easy_install import PthDistributions
 from setuptools.dist import Distribution
 
+from .fixtures import make_sdist
 from .textwrap import DALS
 
 import distutils.errors
@@ -378,186 +375,6 @@ class TestUserInstallTest:
         actual = os.path.normcase(os.path.realpath(res.location))
         expected = os.path.normcase(os.path.realpath(foo_package))
         assert actual == expected
-
-
-def make_trivial_sdist(dist_path, distname, version):
-    """
-    Create a simple sdist tarball at dist_path, containing just a simple
-    setup.py.
-    """
-
-    make_sdist(
-        dist_path,
-        [
-            (
-                'setup.py',
-                DALS(
-                    f"""\
-             import setuptools
-             setuptools.setup(
-                 name={distname!r},
-                 version={version!r}
-             )
-         """
-                ),
-            ),
-            ('setup.cfg', ''),
-        ],
-    )
-
-
-def make_nspkg_sdist(dist_path, distname, version):
-    """
-    Make an sdist tarball with distname and version which also contains one
-    package with the same name as distname.  The top-level package is
-    designated a namespace package).
-    """
-    # Assert that the distname contains at least one period
-    assert '.' in distname
-
-    parts = distname.split('.')
-    nspackage = parts[0]
-
-    packages = ['.'.join(parts[:idx]) for idx in range(1, len(parts) + 1)]
-
-    setup_py = DALS(
-        f"""\
-        import setuptools
-        setuptools.setup(
-            name={distname!r},
-            version={version!r},
-            packages={packages!r},
-            namespace_packages=[{nspackage!r}]
-        )
-    """
-    )
-
-    init = "__import__('pkg_resources').declare_namespace(__name__)"
-
-    files = [('setup.py', setup_py), (os.path.join(nspackage, '__init__.py'), init)]
-    for package in packages[1:]:
-        filename = os.path.join(*(package.split('.') + ['__init__.py']))
-        files.append((filename, ''))
-
-    make_sdist(dist_path, files)
-
-
-def make_python_requires_sdist(dist_path, distname, version, python_requires):
-    make_sdist(
-        dist_path,
-        [
-            (
-                'setup.py',
-                DALS(
-                    """\
-                import setuptools
-                setuptools.setup(
-                  name={name!r},
-                  version={version!r},
-                  python_requires={python_requires!r},
-                )
-                """
-                ).format(
-                    name=distname, version=version, python_requires=python_requires
-                ),
-            ),
-            ('setup.cfg', ''),
-        ],
-    )
-
-
-def make_sdist(dist_path, files):
-    """
-    Create a simple sdist tarball at dist_path, containing the files
-    listed in ``files`` as ``(filename, content)`` tuples.
-    """
-
-    # Distributions with only one file don't play well with pip.
-    assert len(files) > 1
-    with tarfile.open(dist_path, 'w:gz') as dist:
-        for filename, content in files:
-            file_bytes = io.BytesIO(content.encode('utf-8'))
-            file_info = tarfile.TarInfo(name=filename)
-            file_info.size = len(file_bytes.getvalue())
-            file_info.mtime = int(time.time())
-            dist.addfile(file_info, fileobj=file_bytes)
-
-
-def create_setup_requires_package(
-    path,
-    distname='foobar',
-    version='0.1',
-    make_package=make_trivial_sdist,
-    setup_py_template=None,
-    setup_attrs=None,
-    use_setup_cfg=(),
-):
-    """Creates a source tree under path for a trivial test package that has a
-    single requirement in setup_requires--a tarball for that requirement is
-    also created and added to the dependency_links argument.
-
-    ``distname`` and ``version`` refer to the name/version of the package that
-    the test package requires via ``setup_requires``.  The name of the test
-    package itself is just 'test_pkg'.
-    """
-
-    normalized_distname = safer_name(distname)
-    test_setup_attrs = {
-        'name': 'test_pkg',
-        'version': '0.0',
-        'setup_requires': [f'{normalized_distname}=={version}'],
-        'dependency_links': [os.path.abspath(path)],
-    }
-    if setup_attrs:
-        test_setup_attrs.update(setup_attrs)
-
-    test_pkg = os.path.join(path, 'test_pkg')
-    os.mkdir(test_pkg)
-
-    # setup.cfg
-    if use_setup_cfg:
-        options = []
-        metadata = []
-        for name in use_setup_cfg:
-            value = test_setup_attrs.pop(name)
-            if name in 'name version'.split():
-                section = metadata
-            else:
-                section = options
-            if isinstance(value, (tuple, list)):
-                value = ';'.join(value)
-            section.append(f'{name}: {value}')
-        test_setup_cfg_contents = DALS(
-            """
-            [metadata]
-            {metadata}
-            [options]
-            {options}
-            """
-        ).format(
-            options='\n'.join(options),
-            metadata='\n'.join(metadata),
-        )
-    else:
-        test_setup_cfg_contents = ''
-    with open(os.path.join(test_pkg, 'setup.cfg'), 'w', encoding="utf-8") as f:
-        f.write(test_setup_cfg_contents)
-
-    # setup.py
-    if setup_py_template is None:
-        setup_py_template = DALS(
-            """\
-            import setuptools
-            setuptools.setup(**%r)
-        """
-        )
-    with open(os.path.join(test_pkg, 'setup.py'), 'w', encoding="utf-8") as f:
-        f.write(setup_py_template % test_setup_attrs)
-
-    foobar_path = os.path.join(path, f'{normalized_distname}-{version}.tar.gz')
-    make_package(foobar_path, distname, version)
-
-    return test_pkg
 
 
 @pytest.mark.skipif(
